1. Ajustes dos endpoint; 2. Criacao do pkg Utils e insercao de ENUM de erros.
