+++ backend/errorcenter/src/main/java/br/com/codenation/errorcenter/controller/LogController.java
