+++ backend/errorcenter/src/main/java/br/com/codenation/errorcenter/controller/UserController.java
@@ -1,0 +1,27 @@
+package br.com.codenation.errorcenter.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.codenation.errorcenter.models.User;
+import br.com.codenation.errorcenter.service.UserService;
+
+@RestController
+@RequestMapping("/user")
+public class UserController {
+
+	@Autowired
+	private UserService userService;
+
+	@PostMapping
+	public ResponseEntity<?> insertUsers(@RequestBody List<User> user){
+		userService.save(user);
+		return ResponseEntity.ok(user);
+	}
+}
