+++ backend/errorcenter/src/main/java/br/com/codenation/errorcenter/models/Status.java
