+++ backend/errorcenter/src/main/java/br/com/codenation/errorcenter/repository/UserRepository.java
@@ -3,10 +3,12 @@ package br.com.codenation.errorcenter.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import br.com.codenation.errorcenter.models.User;
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long>{
 	
-	public Optional<User> findByName(String name);
+	public Optional<User> findByEmail(String email);
 }
