+++ backend/errorcenter/src/main/java/br/com/codenation/errorcenter/service/LogService.java
